# Improve the performance of z.nativeEnum