--- a/deno/lib/types.ts
+++ b/deno/lib/types.ts
@@ -4462,12 +4462,12 @@ export class ZodNativeEnum<T extends EnumLike> extends ZodType<
   #cache: Set<T[keyof T]> | undefined;
   _parse(input: ParseInput): ParseReturnType<T[keyof T]> {
     const ctx = this._getOrReturnCtx(input);
-    
+
     if (
       ctx.parsedType !== ZodParsedType.string &&
       ctx.parsedType !== ZodParsedType.number
     ) {
-      const expectedValues = util.getValidEnumValues(this._def.values)
+      const expectedValues = util.getValidEnumValues(this._def.values);
       addIssueToContext(ctx, {
         expected: util.joinValues(expectedValues) as "string",
         received: ctx.parsedType,
@@ -4481,7 +4481,7 @@ export class ZodNativeEnum<T extends EnumLike> extends ZodType<
     }
 
     if (!this.#cache.has(input.data)) {
-      const expectedValues = util.getValidEnumValues(this._def.values)
+      const expectedValues = util.getValidEnumValues(this._def.values);
 
       addIssueToContext(ctx, {
         received: ctx.data,