Improve the performance of z.nativeEnum

The previous implementation converted the enum into an array of values
on each parse call, even though the array was never used in the happy
path.

This was especially problematic on large enums, since the work done was
proportional to the size of the enum.

I also removed an extra call to `util.objectValues` on the failure path, since
it seemed to be superfluous (`util.getValidEnumValues` already calls it).

Before:

    z.nativeEnum: valid x 4,376,561 ops/sec ±0.65% (95 runs sampled)
    z.nativeEnum: invalid x 139,663 ops/sec ±6.37% (91 runs sampled)
    long z.nativeEnum: valid x 1,197,450 ops/sec ±0.25% (97 runs sampled)
    long z.nativeEnum: invalid x 116,181 ops/sec ±0.46% (96 runs sampled)

After:

    z.nativeEnum: valid x 10,978,553 ops/sec ±0.41% (98 runs sampled)
    z.nativeEnum: invalid x 143,931 ops/sec ±5.40% (93 runs sampled)
    long z.nativeEnum: valid x 9,783,582 ops/sec ±0.38% (99 runs sampled)
    long z.nativeEnum: invalid x 122,292 ops/sec ±0.48% (99 runs sampled